AzureFunctions/requirements.txt
azure-functions
tweepy
azure-storage-blob
